+++ tool_packages/unique_deep_research/CHANGELOG.md
@@ -5,6 +5,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), 
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.4.0] - 2026-02-20
+- Add option to show full page content of search results in the return of the web search tool
+
 ## [3.3.2] - 2026-02-11
 - Correctly log debug info in case of error